backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
pydantic==2.11.9
pydantic-settings==2.11.0
python-multipart==0.0.20
motor==3.7.1
pymongo==4.15.5
redis==6.4.0
chromadb==1.1.0
lancedb==0.25.2
langchain==1.2.10
langchain-core==1.2.14
langchain-openai==1.1.10
langchain-ollama==1.0.1
langchain-mongodb==0.11.0
langchain-deepseek==1.0.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-checkpoint-mongodb==0.3.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
requests==2.32.5
httpx==0.28.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
openai==2.21.0
ollama==0.6.0